refactor(app): split components/ into layouts/ and feature sections/
Pure git mv + import-specifier rewrite, no logic changes:
- src/layouts/: RootLayout, Layout (AppDataProvider), BareLayout, NavBar,
  Footer, MastheadRule
- src/sections/landing/: HeroSection, NumbersStrip, PaletteStrip,
  LibrariesSection, PlotOfTheDay, PlotOfTheDayTerminal, TypewriterText,
  ScienceNote
- src/sections/spec-detail/: SpecDetailView, SpecTabs, SpecOverview,
  RelatedSpecs, CodeShowcase, LibraryPills (used only by SpecPage — moved
  here instead of plots-gallery as originally sketched)
- src/sections/plots-gallery/: FilterBar, ImagesGrid, ImageCard,
  ToolbarActions
- src/sections/libraries/: LibraryCard
- src/components/ keeps the shared primitives: LoaderSpinner, SectionHeader,
  ErrorBoundary, RouteErrorBoundary, CodeHighlighter, ThemeToggle,
  FeedbackWidget
- Barrel index.ts per folder; hooks/ and utils/ barrels completed
- Tests moved alongside; layouts/ and sections/ added to coverage

Part 7 of the frontend modernization roadmap.
