+++ app/src/components/index.ts
@@ -1,0 +1,7 @@
+export * from 'src/components/CodeHighlighter';
+export * from 'src/components/ErrorBoundary';
+export * from 'src/components/FeedbackWidget';
+export * from 'src/components/LoaderSpinner';
+export * from 'src/components/RouteErrorBoundary';
+export * from 'src/components/SectionHeader';
+export * from 'src/components/ThemeToggle';
