+++ app/src/sections/landing/NumbersStrip.tsx
