+++ app/src/sections/landing/PaletteStrip.tsx
