+++ app/src/sections/landing/PlotOfTheDay.tsx
