+++ app/src/sections/landing/PlotOfTheDayTerminal.tsx
