+++ app/src/sections/landing/TypewriterText.tsx
