+++ app/src/sections/landing/index.ts
@@ -1,0 +1,8 @@
+export * from 'src/sections/landing/HeroSection';
+export * from 'src/sections/landing/LibrariesSection';
+export * from 'src/sections/landing/NumbersStrip';
+export * from 'src/sections/landing/PaletteStrip';
+export * from 'src/sections/landing/PlotOfTheDay';
+export * from 'src/sections/landing/PlotOfTheDayTerminal';
+export * from 'src/sections/landing/ScienceNote';
+export * from 'src/sections/landing/TypewriterText';
