+++ app/src/sections/plots-gallery/ImageCard.tsx
