+++ app/src/sections/plots-gallery/index.ts
@@ -1,0 +1,4 @@
+export * from 'src/sections/plots-gallery/FilterBar';
+export * from 'src/sections/plots-gallery/ImageCard';
+export * from 'src/sections/plots-gallery/ImagesGrid';
+export * from 'src/sections/plots-gallery/ToolbarActions';
