+++ app/src/sections/spec-detail/CodeShowcase.tsx
