+++ app/src/sections/spec-detail/LibraryPills.tsx
