+++ app/src/sections/spec-detail/RelatedSpecs.tsx
