+++ app/src/sections/spec-detail/SpecDetailView.tsx
