+++ app/src/sections/spec-detail/SpecOverview.tsx
