+++ app/src/sections/spec-detail/SpecTabs.tsx
