+++ app/src/sections/spec-detail/index.ts
@@ -1,0 +1,6 @@
+export * from 'src/sections/spec-detail/CodeShowcase';
+export * from 'src/sections/spec-detail/LibraryPills';
+export * from 'src/sections/spec-detail/RelatedSpecs';
+export * from 'src/sections/spec-detail/SpecDetailView';
+export * from 'src/sections/spec-detail/SpecOverview';
+export * from 'src/sections/spec-detail/SpecTabs';
