+++ app/src/utils/index.ts
@@ -5,3 +5,7 @@
 export * from 'src/utils/tooltip';
 export * from 'src/utils/filters';
 export * from 'src/utils/fuzzySearch';
+export * from 'src/utils/claudePrompt';
+export * from 'src/utils/responsiveImage';
+export * from 'src/utils/shuffle';
+export * from 'src/utils/themedPreview';
